+++ src/ezdxf/layouts/layouts.py
@@ -1,7 +1,6 @@
-# Created: 21.03.2011
 # Copyright (c) 2011-2020, Manfred Moitzi
 # License: MIT License
-from typing import TYPE_CHECKING, Dict, Iterable, List, cast, Optional, Set
+from typing import TYPE_CHECKING, Dict, Iterable, List, cast, Optional
 import logging
 from ezdxf.lldxf.const import DXFKeyError, DXFValueError, DXFInternalEzdxfError
 from ezdxf.lldxf.const import (
@@ -18,13 +17,24 @@ if TYPE_CHECKING:
 logger = logging.getLogger('ezdxf')
 
 
+def key(name: str) -> str:
+    """ AutoCAD uses case insensitive layout names, but stores the name case
+    sensitive. """
+    return name.upper()
+
+
+MODEL = key('Model')
+
+
 class Layouts:
     def __init__(self, doc: 'Drawing'):
         """ Default constructor. (internal API) """
         self.doc = doc
+        # Store layout names in normalized form: key(name)
         self._layouts: Dict[str, Layout] = {}
-        # key: layout name; value: DXFLayout()
-        self._dxf_layouts: 'Dictionary' = self.doc.rootdict['ACAD_LAYOUT']
+        # key: layout name as original case sensitive string; value: DXFLayout()
+        self._dxf_layouts: 'Dictionary' = cast('Dictionary',
+                                               self.doc.rootdict['ACAD_LAYOUT'])
 
     @classmethod
     def setup(cls, doc: 'Drawing'):
@@ -33,6 +43,32 @@ class Layouts:
         layouts.setup_modelspace()
         layouts.setup_paperspace()
         return layouts
+
+    def __len__(self) -> int:
+        """ Returns count of existing layouts, including the modelspace
+        layout. """
+        return len(self._layouts)
+
+    def __contains__(self, name: str) -> bool:
+        """ Returns ``True`` if layout `name` exist. """
+        assert isinstance(name, str), type(str)
+        return key(name) in self._layouts
+
+    def __iter__(self) -> Iterable['Layout']:
+        """ Returns iterable of all layouts as :class:`~ezdxf.layouts.Layout`
+        objects, including the modelspace layout.
+        """
+        return iter(self._layouts.values())
+
+    def _add_layout(self, name: str, layout: Layout):
+        layout.dxf.name = name
+        self._layouts[key(name)] = layout
+        self._dxf_layouts[name] = layout.dxf_layout
+
+    def _discard(self, layout: 'Layout'):
+        name = layout.name
+        self._dxf_layouts.discard(name)
+        del self._layouts[key(name)]
 
     def setup_modelspace(self):
         """ Modelspace setup. (internal API) """
@@ -50,9 +86,7 @@ class Layouts:
             raise DXFValueError(f'Layout "{name}" already exists')
         dxfattribs['owner'] = self._dxf_layouts.dxf.handle
         layout = cls.new(name, block_name, self.doc, dxfattribs=dxfattribs)
-        self._dxf_layouts[name] = layout.dxf_layout
-        self._layouts[name] = layout
-
+        self._add_layout(name, layout)
         return layout
 
     def unique_paperspace_name(self) -> str:
@@ -80,8 +114,8 @@ class Layouts:
         if not is_valid_table_name(name):
             raise DXFValueError('Layout name contains invalid characters.')
 
-        if name.upper() in self.upper_names():
-            raise DXFValueError(f'Layout "{name}" already exists')
+        if name in self:
+            raise DXFValueError(f'Layout "{name}" already exist.')
 
         dxfattribs = dict(dxfattribs or {})  # copy attribs
         dxfattribs['owner'] = self._dxf_layouts.dxf.handle
@@ -89,10 +123,7 @@ class Layouts:
         block_name = self.unique_paperspace_name()
         layout = Paperspace.new(name, block_name, self.doc,
                                 dxfattribs=dxfattribs)
-
-        self._dxf_layouts[name] = layout.dxf_layout
-        self._layouts[name] = layout
-
+        self._add_layout(name, layout)
         return layout
 
     @classmethod
@@ -109,9 +140,9 @@ class Layouts:
         return layouts
 
     def restore(self, name: str, block_record_name: str, taborder: int) -> None:
-        """ Restore layout from block if DXFLayout do not exist.
+        """ Restore layout from block if DXFLayout does not exist.
         (internal API) """
-        if name in self._layouts:
+        if name in self:
             return
         block_layout = self.doc.blocks.get(block_record_name)
         self._new_from_block_layout(name, block_layout, taborder)
@@ -126,64 +157,42 @@ class Layouts:
         }
         dxf_layout = cast('DXFLayout', self.doc.objects.new_entity(
             'LAYOUT', dxfattribs=dxfattribs))
-        if name == 'Model':
+        if key(name) == MODEL:
             layout = Modelspace.load(dxf_layout, self.doc)
         else:
             layout = Paperspace.load(dxf_layout, self.doc)
-
-        self._dxf_layouts[name] = layout.dxf_layout
-        self._layouts[name] = layout
+        self._add_layout(name, layout)
         return layout
 
     def setup_from_rootdict(self) -> None:
         """ Setup layout manger from root dictionary. (internal API) """
         for name, dxf_layout in self._dxf_layouts.items():
-            if name.upper() == 'MODEL':
+            if key(name) == MODEL:
                 layout = Modelspace(dxf_layout, self.doc)
-                name = 'Model'
             else:
                 layout = Paperspace(dxf_layout, self.doc)
-            self._layouts[name] = layout
-
-    def __len__(self) -> int:
-        """ Returns count of existing layouts, including the modelspace
-        layout. """
-        return len(self._layouts)
-
-    def __contains__(self, name: str) -> bool:
-        """ Returns ``True`` if layout `name` exist. """
-        assert isinstance(name, str), type(str)
-        return name in self._layouts
-
-    def __iter__(self) -> Iterable['Layout']:
-        """ Returns iterable of all layouts as :class:`~ezdxf.layouts.Layout`
-        objects, including the modelspace layout.
-        """
-        return iter(self._layouts.values())
+            # assert name == layout.dxf.name
+            self._layouts[key(name)] = layout
 
     def modelspace(self) -> Modelspace:
         """ Returns the :class:`~ezdxf.layouts.Modelspace` layout. """
         return cast(Modelspace, self.get('Model'))
 
-    def names(self) -> Iterable[str]:
-        """ Returns iterable of all layout names. """
-        return self._layouts.keys()
-
-    def upper_names(self) -> Set[str]:
-        return {name.upper() for name in self.names()}
+    def names(self) -> List[str]:
+        """ Returns a list of all layout names, all names in original case
+        sensitive form. """
+        return [layout.name for layout in self._layouts.values()]
 
     def get(self, name: Optional[str]) -> 'Layout':
-        """ Returns :class:`~ezdxf.layouts.Layout` by `name`.
+        """ Returns :class:`~ezdxf.layouts.Layout` by `name`, case insensitive
+        "Model" == "MODEL".
 
         Args:
             name: layout name as shown in tab, e.g. ``'Model'`` for modelspace
 
         """
-        if name is None:
-            first_layout_name = self.names_in_taborder()[1]
-            return self._layouts[first_layout_name]
-        else:
-            return self._layouts[name]
+        name = name or self.names_in_taborder()[1]  # first paperspace layout
+        return self._layouts[key(name)]
 
     def rename(self, old_name: str, new_name: str) -> None:
         """ Rename a layout from `old_name` to `new_name`.
@@ -191,8 +200,8 @@ class Layouts:
         not exist.
 
         Args:
-            old_name: actual layout name
-            new_name: new layout name
+            old_name: actual layout name, case insensitive
+            new_name: new layout name, case insensitive
 
         Raises:
             DXFValueError: try to rename ``'Model'``
@@ -201,21 +210,23 @@ class Layouts:
         """
         assert isinstance(old_name, str), type(old_name)
         assert isinstance(new_name, str), type(new_name)
-        if old_name.upper() == 'MODEL':
+        if key(old_name) == MODEL:
             raise DXFValueError('Can not rename model space.')
-        if new_name.upper() in self.upper_names():
+        if new_name in self:
             raise DXFValueError(f'Layout "{new_name}" already exist.')
+        if old_name not in self:
+            raise DXFValueError(f'Layout "{old_name}" does not exist.')
 
-        layout = self._layouts[old_name]
+        layout = self.get(old_name)
+        self._discard(layout)
         layout.rename(new_name)
-        del self._layouts[old_name]
-        self._layouts[new_name] = layout
+        self._add_layout(new_name, layout)
 
     def names_in_taborder(self) -> List[str]:
         """ Returns all layout names in tab order as shown in :term:`CAD`
         applications. """
-        names = [(layout.dxf.taborder, name) for name, layout in
-                 self._layouts.items()]
+        names = [(layout.dxf.taborder, layout.name) for layout in
+                 self._layouts.values()]
         return [name for order, name in sorted(names)]
 
     def get_layout_for_entity(self, entity: 'DXFEntity') -> 'Layout':
@@ -245,7 +256,7 @@ class Layouts:
     def set_active_layout(self, name: str) -> None:
         """ Set layout `name` as active paperspace layout. """
         assert isinstance(name, str), type(name)
-        if name.upper() == 'MODEL':  # reserved layout name
+        if key(name) == MODEL:  # reserved layout name
             raise DXFValueError('Can not set model space as active layout')
         # raises KeyError if layout 'name' does not exist
         new_active_layout = self.get(name)
@@ -261,7 +272,7 @@ class Layouts:
         blocks.rename_block(TMP_PAPER_SPACE_NAME, new_active_paper_space_name)
 
     def delete(self, name: str) -> None:
-        """ Delete layout `name` and all entities owned by it.
+        """ Delete layout `name` and destroy all entities in that layout.
 
         Args:
             name (str): layout name as shown in tabs
@@ -273,21 +284,20 @@ class Layouts:
 
         """
         assert isinstance(name, str), type(name)
-        if name.upper() == 'MODEL':
+        if key(name) == MODEL:
             raise DXFValueError("Can not delete modelspace layout.")
 
-        layout = self._layouts[name]
+        layout = self.get(name)
         if len(self) < 3:
             raise DXFValueError("Can not delete last paperspace layout.")
         if layout.layout_key == self.get_active_layout_key():
             # Layout `name` is the active layout:
-            for layout_name in self.names():
+            for layout_name in self._layouts:
                 # Set any other paperspace layout as active layout
-                if layout_name not in (name, 'Model'):
+                if layout_name not in (key(name), MODEL):
                     self.set_active_layout(layout_name)
                     break
-        self._dxf_layouts.remove(layout.name)
-        del self._layouts[layout.name]
+        self._discard(layout)
         layout.destroy()
 
     def active_layout(self) -> Paperspace:
@@ -319,7 +329,7 @@ class Layouts:
         }
         psp_layout_br_handles = {
             layout.dxf.block_record_handle for layout in
-            self._layouts.values() if layout.dxf.name.upper() != 'MODEL'
+            self._layouts.values() if key(layout.name) != MODEL
         }
         mismatch = psp_br_handles.difference(psp_layout_br_handles)
         if len(mismatch):